[Template] Consolidate menu-popup and offer list for incito and paged
