+++ examples/list-publications-template.html
@@ -27,7 +27,7 @@
             id="sgn-sdk"
             data-api-key="GL0wU8"
             data-track-id="2N2Gib"
-            data-business-id="44d0Q"
+            data-business-id="c1edq"
             data-component="list-publications"
             data-component-list-publications-container="#list-publications"
             data-component-publications-viewer-preferred-type="incito"
