+++ lib/kits/core-ui/components/common/header.js
@@ -86,14 +86,14 @@ const defaultShoppingListCounterTemplate = `\
 `;
 
 const Header = ({
-    type,
+    publicationType,
     template,
     shoppingListCounterTemplate,
     el,
     scriptEls
 }) => {
     let container = null;
-    type = type || 'paged';
+    publicationType = publicationType || 'paged';
     template = template?.innerHTML || defaultTemplate;
 
     const renderShoppingListCounter = () => {
@@ -117,8 +117,6 @@ const Header = ({
     const setNavColor = (color) => {
         const sgnNav = container.querySelector('.sgn__nav');
 
-        console.log(color, getColorBrightness(color));
-
         if (sgnNav) {
             sgnNav.style.backgroundColor = color || 'transparent';
             sgnNav.style.color =
@@ -140,7 +138,7 @@ const Header = ({
 
     const addClosePubListener = () => {
         const sgnContainer =
-            type === 'incito'
+            publicationType === 'incito'
                 ? el.querySelector('.sgn__incito')
                 : el.querySelector('.sgn__pp');
         const closeBtn = container.querySelector('.sgn__close-publication');
@@ -210,10 +208,10 @@ const Header = ({
                 scriptEls.offerClickBehavior !== 'shopping_list',
             disableMenu: scriptEls.disableMenu,
             disableClose: scriptEls.disableClose,
-            isIncito: type === 'incito'
+            isIncito: publicationType === 'incito'
         });
 
-        if (type === 'incito') {
+        if (publicationType === 'incito') {
             addScrollListener();
         }
