+++ lib/kits/core-ui/components/common/menu-popup.js
@@ -1,9 +1,10 @@
 import Mustache from 'mustache';
 import {createModal, formatDate, translate} from '../helpers/component';
-import './menu-popup.styl';
 import OfferList from './offer-list';
 import PageList from '../paged-publication/page-list';
+import SectionList from '../incito-publication/section-list';
 import PublicationDownload from './publication-download';
+import './menu-popup.styl';
 
 const defaultTemplate = `\
     <div class="sgn-popup-container sgn-menu-popup">
@@ -21,26 +22,61 @@ const defaultTemplate = `\
             </div>
         {{/disableDownload}}
         <div class="sgn-clearfix"></div>
-        <div class="sgn-menu-tab">
-            <button class="sgn-menu-tab-btn sgn-menu-tab-btn-active">{{translations.pagesButton}}</button>
-            <button class="sgn-menu-tab-btn">{{translations.offersButton}}</button>
-        </div>
-        <div class="sgn-menu-tab-content-container">
-            <div class="sgn-menu-tab-content sgn-menu-tab-content-active">
-                <div class="sgn-menu-tab-content-pages"></div>
+        {{#isIncito}}
+            {{#hasSections}}
+            <div class="sgn-menu-tab">
+                <button class="sgn-menu-tab-btn sgn-menu-tab-btn-active">{{translations.overviewButton}}</button>
+                <button class="sgn-menu-tab-btn">{{translations.offersButton}}</button>
             </div>
-            <div class="sgn-menu-tab-content">
-                <div class="sgn-menu-tab-content-offers"></div>
+            <div class="sgn-menu-tab-content-container">
+                <div class="sgn-menu-tab-content sgn-menu-tab-content-active">
+                    <div class="sgn-menu-tab-content-sections"></div>
+                </div>
+                <div class="sgn-menu-tab-content">
+                    <div class="sgn-menu-tab-content-offers"></div>
+                </div>
             </div>
+            {{/hasSections}}
+            {{^hasSections}}
+            <div class="sgn-menu-tab-content-container">
+                <div class="sgn-menu-tab-content sgn-menu-tab-content-active">
+                    <div class="sgn-menu-tab-content-offers"></div>
+                </div>
+            </div>
+            {{/hasSections}}
+        {{/isIncito}}
+        {{^isIncito}}
+            <div class="sgn-menu-tab">
+                <button class="sgn-menu-tab-btn sgn-menu-tab-btn-active">{{translations.pagesButton}}</button>
+                <button class="sgn-menu-tab-btn">{{translations.offersButton}}</button>
+            </div>
+            <div class="sgn-menu-tab-content-container">
+                <div class="sgn-menu-tab-content sgn-menu-tab-content-active">
+                    <div class="sgn-menu-tab-content-pages"></div>
+                </div>
+                <div class="sgn-menu-tab-content">
+                    <div class="sgn-menu-tab-content-offers"></div>
+                </div>
         </div>
+        {{/isIncito}}
     </div>\
 `;
 
-const MenuPopup = ({configs, sgnData, sgnViewer, templates, scriptEls}) => {
+const MenuPopup = ({
+    publicationType,
+    configs,
+    sgnData,
+    sgnViewer,
+    templates,
+    scriptEls
+}) => {
+    publicationType = publicationType || 'paged';
+
     const translations = {
         localeCode: translate('locale_code'),
         untilLabel: translate('publication_viewer_until_label'),
         pagesButton: translate('publication_viewer_pages_button'),
+        overviewButton: translate('publication_viewer_overview_button'),
         offersButton: translate('publication_viewer_offers_button'),
         downloadButton: translate('publication_viewer_download_button')
     };
@@ -67,12 +103,15 @@ const MenuPopup = ({configs, sgnData, sgnViewer, templates, scriptEls}) => {
                 ),
                 disableDownload: sgnData?.details?.pdf_url
                     ? scriptEls.disableDownload
-                    : true
+                    : true,
+                isIncito: publicationType === 'incito',
+                hasSections: sgnData?.incito?.table_of_contents?.length > 0
             }
         );
 
         createModal(container);
         renderPageList();
+        renderSectionList();
         renderOfferList();
         addTabClickListener();
         renderDownload();
@@ -117,7 +156,7 @@ const MenuPopup = ({configs, sgnData, sgnViewer, templates, scriptEls}) => {
     };
 
     const renderPageList = async () =>
-        container.querySelector('.sgn-menu-tab-content-pages').appendChild(
+        container.querySelector('.sgn-menu-tab-content-pages')?.appendChild(
             await PageList({
                 configs,
                 sgnData,
@@ -126,9 +165,18 @@ const MenuPopup = ({configs, sgnData, sgnViewer, templates, scriptEls}) => {
             }).render()
         );
 
+    const renderSectionList = async () =>
+        container.querySelector('.sgn-menu-tab-content-sections')?.appendChild(
+            await SectionList({
+                sgnData,
+                template: templates.sectionList
+            }).render()
+        );
+
     const renderOfferList = async () =>
         container.querySelector('.sgn-menu-tab-content-offers').appendChild(
             await OfferList({
+                publicationType,
                 configs,
                 sgnViewer,
                 template: templates.offerList
