+++ lib/kits/core-ui/components/common/menu-popup.styl
@@ -115,3 +115,19 @@
         a
             color #F3F3F9
 
+.sgn-modal-container
+    position fixed
+    top 0
+    left 0
+    bottom 0
+    right 0
+    content ' '
+    background rgba(0, 0, 0, 0.1)
+
+.sgn-blocker
+    position fixed
+    top 0
+    left 0
+    bottom 0
+    right 0
+    content ' '
