+++ lib/kits/core-ui/components/common/offer-list.js
@@ -11,7 +11,7 @@ const defaultTemplate = `\
         <ol class="sgn-offers-list-items-container">
             {{#offers}}
             <li class="sgn-offers-list-item-container">
-                <div class="sgn-offers-content-container" data-page-id="{{pageId}}">
+                <div class="sgn-offers-content-container" data-page-id="{{pageId}}" data-view-id="{{catalog_view_id}}">
                     <div class="sgn-offers-content-img">
                         <img src="{{images.view}}" alt="{{heading}}">
                     </div>
@@ -34,8 +34,9 @@ const defaultTemplate = `\
     </div>\
 `;
 
-const OfferList = ({configs = {}, sgnViewer, template}) => {
+const OfferList = ({publicationType, configs = {}, sgnViewer, template}) => {
     template = template?.innerHTML || defaultTemplate;
+    publicationType = publicationType || 'paged';
 
     const offerTemplate = template.match(
         /(?={{#offers}}).*({{\/offers}})/gs
@@ -62,6 +63,7 @@ const OfferList = ({configs = {}, sgnViewer, template}) => {
                       qs: {
                           dealer_id: configs.businessId,
                           catalog_id: configs.id,
+                          types: publicationType,
                           limit,
                           query
                       }
@@ -78,6 +80,7 @@ const OfferList = ({configs = {}, sgnViewer, template}) => {
                 qs: {
                     dealer_id: configs.businessId,
                     catalog_id: configs.id,
+                    types: publicationType,
                     offset,
                     limit
                 }
@@ -148,13 +151,29 @@ const OfferList = ({configs = {}, sgnViewer, template}) => {
         container
             .querySelectorAll('.sgn-offers-content-container')
             .forEach((itemEl) => {
-                itemEl.addEventListener('click', redirectToPage, false);
+                itemEl.addEventListener(
+                    'click',
+                    publicationType === 'incito'
+                        ? scrollToOffer
+                        : redirectToPage,
+                    false
+                );
             });
     };
 
     const redirectToPage = (e) => {
         destroyModal();
         sgnViewer.navigateToPageId(e.currentTarget.dataset?.pageId);
+    };
+
+    const scrollToOffer = (e) => {
+        const viewid = e.currentTarget.dataset?.viewId;
+        const offerCell = document.querySelector(
+            `[data-id="${viewid}"][data-role="offer"]`
+        );
+
+        destroyModal();
+        offerCell.scrollIntoView({behavior: 'smooth'});
     };
 
     const render = async () => {
