+++ lib/kits/core-ui/components/incito-publication/index.js
@@ -1,4 +1,2 @@
 export {default as MainContainer} from './main-container';
-export {default as MenuPopup} from './menu-popup';
-export {default as OfferList} from './offer-list';
 export {default as SectionList} from './section-list';
