--- lib/kits/core-ui/components/incito-publication/menu-popup.styl
@@ -1,117 +1,0 @@
-.sgn-menu-popup
-    position relative
-    background #F3F3F9
-    color #202020
-    border-radius 12px
-    padding 10px
-    width 90%
-    max-width 666px
-    min-height 200px
-    margin auto
-    margin-top 80px
-    box-shadow rgba(0, 0, 0, 0.7) 0px 20px 55px, rgba(0, 0, 0, 0.3) 0px 0px 1px
-    text-align left
-
-.sgn-menu-popup-labels
-    padding 10px 0px
-    line-height 1.5
-    float left
-    width 80%
-
-.sgn-menu-download
-    text-align right
-    float left
-    width 20%
-    padding 10px 0px
-
-    > button
-        background-color #d3d3d3
-        padding 4px 10px
-        font-size 12px
-        border-radius 6px
-        border 0
-        transition .2s
-        
-        &:hover
-            opacity 0.8
-            cursor pointer
-
-.sgn-menu-label
-    font-weight 600
-    font-size 18px
-
-.sgn-menu-till
-    font-weight 400
-    font-size 16px
-    vertical-align baseline
-    border-radius 0.25em
-    color rgb(40, 175, 76)
-
-.sgn-menu-tab
-    color #202020
-    background-color rgba(139, 140, 143, 0.1)
-    overflow hidden
-    font-size 12px
-    border-radius 8px
-    align-items center
-
-
-.sgn-menu-tab-btn
-    background-color #d3d3d3
-    color #202020
-    font-size 12px
-    float left
-    border none
-    outline none
-    cursor pointer
-    padding 8px 3px
-    transition .3s
-    width 50%
-
-    &:hover
-        font-weight 600
-        background-color transparent
-
-.sgn-menu-tab-btn-active
-    font-weight 600
-    background-color transparent
-
-.sgn-menu-tab-content
-    display none
-
-    a
-        text-decoration none
-        color #202020
-
-.sgn-menu-tab-content-active
-    display block
-
-.sgn__theme-dark
-    .sgn-menu-popup
-        background rgb(28, 28, 30)
-        color #F3F3F9
-        box-shadow rgba(0, 0, 0, 0.7) 0px 20px 55px, rgba(0, 0, 0, 0.3) 0px 0px 1px
-
-    .sgn-menu-download
-        > button
-            color #F3F3F9
-            background-color #2c2c2e
-
-    .sgn-menu-tab
-        color #F3F3F9
-        background-color rgba(139, 140, 143, 0.1)
-    
-    .sgn-menu-tab-btn
-        background-color rgb(44, 44, 46)
-        color #F3F3F9
-
-        &:hover
-            background-color transparent
-    
-    .sgn-menu-tab-btn-active
-        background-color transparent
-
-    .sgn-menu-tab-content
-        a
-            color #F3F3F9
-
