--- lib/kits/core-ui/components/incito-publication/offer-list.styl
@@ -1,115 +1,0 @@
-.sgn-offers-list-items-container
-    height 650px
-    overflow-y scroll
-    list-style-type none
-    padding 0 10px
-    color #000000
-
-.sgn-offers-search-container
-    padding-top 10px
-
-.sgn-offers-search-text
-    border-radius 4px
-    background-color #d3d3d3
-    color #202020
-    border 1px solid rgba(255, 255, 255, 0.15)
-    width 100%
-    box-sizing border-box
-    padding 8px
-
-.sgn-offers-list-item-container
-    width 31%
-    text-align center
-    margin 1%
-    margin-bottom 12px
-    float left
-
-.sgn-offers-list-item-container:nth-child(3n+1)
-    clear left
-
-.sgn-offers-content-container
-    text-decoration none
-    padding 12px
-    box-shadow rgba(0, 0, 0, .3) 0px 0px 10px
-    border-radius 10px
-
-    &:hover
-        cursor pointer
-
-.sgn-offers-content-img
-    background-color rgb(204, 204, 204)
-    background-size cover
-    background-position center center
-    border-radius 6px
-    overflow hidden
-    position relative
-    border 0px
-    width 100%
-    height auto
-    padding-top 100%
-
-    > img    
-        object-fit cover
-        object-position center center
-        width 100%
-        height 100%
-        position absolute
-        inset 0px
-
-.sgn-offers-content-text
-    white-space nowrap
-    overflow hidden
-    color #202020
-    padding 4px 0px 0px 0px
-    text-align left
-
-.sgn-offers-content-heading
-    font-weight 600
-    line-height 1.2
-    white-space nowrap
-    overflow hidden
-    text-overflow ellipsis
-    max-width 300px
-
-.sgn-offers-content-description
-    font-size 0.9em
-    opacity 0.6
-    color inherit
-
-.sgn-offers-content-price
-    font-variant-numeric tabular-nums
-    color rgb(40, 175, 76)
-    font-size 0.9em
-
-.sgn__theme-dark
-    .sgn-offers-search-text
-        background-color rgb(44, 44, 46)
-        color rgb(243, 243, 249)
-        border 2px solid rgba(0, 0, 0, 0.15)
-    
-    .sgn-offers-content-container
-        box-shadow rgba(243, 243, 249, .15) 0px 0px 10px
-    
-    .sgn-offers-content-img
-        background-color rgb(204, 204, 204)
-    
-    .sgn-offers-content-text
-        color #F3F3F9
-
-@media (max-width 450px)
-    .sgn-offers-list-item-container
-        width 48%
-
-    .sgn-offers-list-item-container:nth-child(3n+1)
-        clear none
-
-    .sgn-offers-list-item-container:nth-child(2n+1)
-        clear left
-
-@media (max-height 900px)
-    .sgn-offers-list-items-container
-        height 500px
-
-@media (max-height 790px)
-    .sgn-offers-list-items-container
-        height 360px
