+++ lib/kits/core-ui/components/paged-publication/main-container.styl
@@ -11,20 +11,3 @@
 
 .sgn-clearfix
     clear both
-
-.sgn-modal-container
-    position fixed
-    top 0
-    left 0
-    bottom 0
-    right 0
-    content ' '
-    background rgba(0, 0, 0, 0.1)
-
-.sgn-blocker
-    position fixed
-    top 0
-    left 0
-    bottom 0
-    right 0
-    content ' '
