+++ lib/kits/core-ui/incito-publication.js
@@ -3,7 +3,7 @@ import {getQueryParam} from '../../util';
 import request from '../core/request';
 import Bootstrapper from '../incito-publication/bootstrapper';
 import MainContainer from './components/incito-publication/main-container';
-import MenuPopup from './components/incito-publication/menu-popup';
+import MenuPopup from './components/common/menu-popup';
 import ShoppingList from './components/common/shopping-list';
 import Header from './components/common/header';
 import {transformScriptData} from './components/helpers/transformers';
@@ -50,7 +50,7 @@ const IncitoPublication = (
     }).render();
 
     const header = Header({
-        type: 'incito',
+        publicationType: 'incito',
         template: customTemplates.headerContainer,
         shoppingListCounterTemplate: customTemplates.shoppingListCounter,
         el: document.querySelector(scriptEls.mainContainer),
@@ -79,6 +79,7 @@ const IncitoPublication = (
             ?.addEventListener('click', (e) => {
                 e.preventDefault();
                 MenuPopup({
+                    publicationType: 'incito',
                     configs: options,
                     sgnData,
                     templates: customTemplates,
