+++ lib/kits/core-ui/paged-publication.js
@@ -51,7 +51,7 @@ const PagedPublication = (
     }).render();
 
     const header = Header({
-        type: 'paged',
+        publicationType: 'paged',
         template: customTemplates.headerContainer,
         shoppingListCounterTemplate: customTemplates.shoppingListCounter,
         el: document.querySelector(scriptEls.mainContainer),
@@ -80,6 +80,7 @@ const PagedPublication = (
             ?.addEventListener('click', (e) => {
                 e.preventDefault();
                 MenuPopup({
+                    publicationType: 'paged',
                     configs: options,
                     sgnData,
                     sgnViewer,
